First draft of current monitor changes and motor board updates
This version will change current reading to milliamps instead of pin readings. We will keep the pin readings because that information is valuable. Ultimately, we will need to upgrade JMRI if we change the place in the code where <c> reports back the pin reading to give a percentage of  max current.