Update platformio.ini
Add src_dir parameter to the [platformio] section of platformio.ini. This allows us to change the source folder from "src" to "DCCppEX" so that the project can be compiled in the Arudino IDE as well as PlatformIO. Arduino requires that the folder name where the code resides must be the same as the name of the "main" program. In this case, the program is DCCppEX.ino