Frightrisk current mod (#12)
* deleted files from old folder

* major changes to  current monitor and added motor boards and configuration parameters.

This version will change current reading to milliamps instead of pin readings. We will keep the pin readings because that information is valuable. Ultimately, we will need to upgrade JMRI if we change the place in the code where <c> reports back the pin reading to give a percentage of  max current.

* Add functions functions to return current, triggerMilliamps and maxMilliamps. Change the serial output of the <c> command. This will require changes to JMRI

* minor correction of typos in SerialCommand.cpp