Correct mentions of SHOW_CONFIG_PIN
the SHOW_CONFIG_PIN was mentioned, but not used. Pin A5 was hard coded as the pin that needed to be used to show detailed configuration information at boot. Also, the comments incorrectly said that the pin needed to be connected HIGH. The pin needs to be connected to LOW (ground) for this to work.